+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -6,6 +6,12 @@
 * Added the `webdriver.Capabilities` class.
 * Removed the deprecated `webdriver.Deferred#resolve` and
     `webdriver.promise.resolved` functions.
+* Added native support for the ChromeDriver. Use `selenium-webdriver/chrome`.
+    When using the `Builder`, requesting chrome without specifying a remote
+    server URL will default to the native ChromeDriver implementation.
+* Changed signature of `SeleniumServer` to `SeleniumServer(jar, options)`.
+* Removed the ability to connect to an existing session from the Builder. This
+    feature is intended for use with the browser-based client.
 
 ## v2.33.0